Refactor Entry component to use typescript
