+++ packages/app/src/app/index.d.ts
@@ -1,0 +1,7 @@
+declare global {
+  interface Window {
+    __isTouch: boolean;
+  }
+}
+
+export {};
